Move clipboard code to its dedicated handler
